requirements.txt
torch
transformers
datasets
pymupdf
pillow
